requirements.txt
chainlit~=1.0.200
dataclasses~=0.6
openai~=1.12.0
python-dotenv~=1.0.1
pandas~=2.2.0
langchain~=0.1.7
deepeval~=0.20.65
langchain_google_vertexai~=0.0.5